requirements.txt
asgiref==3.2.5
bandit==1.6.2
coverage==5.0.3
dj-database-url==0.5.0
Django==3.0.4
django-filter==2.2.0
gitdb==4.0.2
GitPython==3.1.0
pbr==5.4.4
psycopg2==2.8.4
pyflakes==2.1.1
python-dotenv==0.12.0
pytz==2019.3
PyYAML==5.3
six==1.14.0
smmap==3.0.1
sqlparse==0.3.1
stevedore==1.32.0
